Organize project package and update bot entry
