+++ DEVLOG.md
@@ -1,0 +1,20 @@
+# Dev Log
+
+Это файл для заметок о разработке бота и ведения логов.
+
+## Логи приложения
+Бот пишет события через модуль `logging`. Уровень логирования задаётся в `main.py`,
+по умолчанию это `INFO`. Логи выводятся в стандартный поток и могут
+перенаправляться в файл при запуске приложения.
+
+## Как просматривать
+При запуске через `python -m telegram_accounting_bot.main` вывод логов
+появляется в консоли. Чтобы сохранить их в файл, можно использовать
+перенаправление:
+
+```bash
+python -m telegram_accounting_bot.main > bot.log 2>&1
+```
+
+Файл `bot.log` затем можно анализировать для поиска ошибок или статистики.
+
