+++ telegram_accounting_bot/__init__.py
@@ -1,0 +1,2 @@
+"""Telegram Accounting Bot package"""
+
